[ffxutil] Increase outer timeout around ffx daemon stop.
Sometimes the daemon fails to stop because it hits the context deadline
first. However, uploading the CAS outputs also then fails probably
because some cleanup is still happening in the background so files that
it tries to map to the outputs end up not existing when it actually
tries to upload. Temporarily extend the outer timeout so that hopefully
the bot won't fail to upload the outputs and we can see the logs to
diagnose what's happening.

Bug: 127370
Change-Id: I20b4e86a4bdfb57b3de0be12c9ba9392b83f4333
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857788
Reviewed-by: Andrés Villegas <andresvi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>