[reference_doc] Support "rename" attribute
The semantics of this attribute is similar to that of serde.

This is used to rename "stop_timeout_ms" to "__stop_timeout_ms" when
rendered to documentation. It looks like we use "__stop_timeout_ms" in
CML in production instead of "stop_timeout_ms".

Change-Id: Idb19b71d8332a53a1dd2ebfc312b5f8831a75847
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857636
Reviewed-by: Adam Perry <adamperry@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>