[ffx][emu] Add instance name default to default config
This adds the default value for `emu.name` to the data file
which is used as the default configuration for ffx.

This is done as a form of documentation, so all the values
which can be configured are printed when running `ffx config get`.

Change-Id: I1c11c933f29df76045cbb7e6a7c88d7b093933fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858101
Fuchsia-Auto-Submit: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>