[stdcompat] Fix std::erase template deduction error
https://reviews.llvm.org/D149543 changed headers such that
vector now transitively includes string. This caused std::string's
std::erase to become a candidate for overload resolution of
std::erase. This was using the template directly and not relying
on argument deduction to weed the std::erase for string out. This
substitution was an error outside of the immediate context of
std::erase so it was not subject to SFINAE.

Instead, just use the left hand side of the assignment to do normal
template argument deduction to choose the correct std::erase.

Fixed: 127378
Change-Id: I214c66d913cc12e6dc8b6ae2de614dbcdc9f8565
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858236
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Alex Brachet <abrachet@google.com>