[starnix] Avoid code injection on module changes
In restricted executor, since the Linux thread and the starnix kernel
are running in the same Zircon thread, it's possible to issue a software
breakpoint in the starnix kernel directly to notify a module change.
This uses less hack and results in a cleaner code.

Change-Id: I2f3d096be5ecdbcdda553c38e765327c2f8eca61
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858017
Commit-Queue: Dangyi Liu <dangyi@google.com>
Reviewed-by: James Robinson <jamesr@google.com>