[debugger] Improve error message when downloads fail
Explicitly print a not found error in the case where a BuildID is not
found on any of the currently configured symbol servers (as opposed to
other types of downloading errors).

Point users to the ZXDB bug template if their symbols are not found or
otherwise not working. With the new information, we should be able to
help debug better.

Fixed: 126624
Change-Id: Iab19f45630fc79faebffaf3f8b9e2d8b4cc72506
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857422
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>