[driver_manager] Specify collection name to runner directly
Because package name is not the only consideration when choosing which
collection to place the driver, we shouldn't rely on using the package
type to determine the collection name in the runner code. This was
missed by the last attempt to remedy this.

Change-Id: Idef2ce20a79035b2e29fd2895d5004a550c7922f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857977
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>