[cf] Test RealmBuilder replacement of existing Config values
The "replace all values" tests added in https://fxrev.dev/695156
initialize the Config to empty and thus are not really replacing values.
Rename these tests and add new tests that initialize the Config with the
values in the package and replace all those values.

Bug: 98743
Change-Id: I51da44fa406247cb071ea8432b3dc27c04968747
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858414
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>