[userspace fastboot] add `flash gpt-meta` support
We're adding support for flashing the gpt-meta pseudo-partition to make
it easy to initialize partition tables.

This triggers the exact same logic as `oem init-partition-tables`, but
aliasing it behind a `flash` command allows us to use existing ffx and
flash manifest features without needing to add one-off support for a
custom oem command.

This is only being implemented for userspace fastboot so far, as the
driving use case is to allow `ffx flash` to completely bootstrap a
device from a USB installer. If we ever need this in Gigaboot fastboot
as well, the trick is going to be making sure both the paver and
Gigaboot come up with the exact same partition tables from this command.

Bug: b/282797139
Change-Id: Ifd671fc9e07669d2fc29773d3bb4eb8d17932b5a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859819
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: David Pursell <dpursell@google.com>