[starnix] Show correct mount root in mountinfo
The 4th column in mountinfo is supposed to be the path from the FS
root to the root of the mount (i.e. it shouldn't depend on whether the
same FS is mounted elsewhere). Starnix was returning an empty string
for non-bind mounts. For bind mounts it was the path to the source
dir in the original mount.

Also removed `Mount.origin_mount` as it's not longer needed.

Bug: 79302
Change-Id: Icd580c1081844dda78d3aa76f389a4fe70d57b00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858340
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Sergey Ulanov <sergeyu@google.com>