[tests][input] Try to de-flake text input test
The text input test has a long-standing history of flaking at
the point where we expect the text input field to receive
focus.

This change tests the hypothesis from fxb/125115 that we
sometimes miss the change of focus message notification
which is passed through the document title in the test
web app.

Assuming that we sometimes miss the events, concatenating
new states into `document.title` instead of replacing
`document.title` with new states, and checking that the
title contains, instead of is-equal-to the required state
may allow us to eliminate flakes.

Bug: 125115
Change-Id: I187f7bfe4d2188abcecf3375069b5ccbcd0d2437
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859792
Reviewed-by: Caroline Liu <carolineliu@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>