[ffx] fix bootstrap flashing order
The bootstrap images should always be flashed first before the
bootloaders, since bootstrap is the lowest-level set of images that the
others may depend on.

For example, bootstrap images might contain the GPT which has to be
flashed first so that the bootloaders will go to the right location on
disk.

Bug: b/282797139
Change-Id: I5bc71e71e9fc044846633dc5165eff1f8da5d0a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859044
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: David Pursell <dpursell@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>