[component][fuzzing] Fix integer overflow
`RealmFuzzerRunner::FuzzInputs` tried to adjust the total number of
configured runs by substracting the runs involving the fixed corpus
elements. Unfortunately, this meant that if the configured runs was low,
but the corpus was large, it would overflow to a large positive value
and run indefinitely. This CL modifies FuzzInputs to always run at least
the fixed corpus inputs and avoid the overflow.

Change-Id: I7e4cabd77f776fc112515e0f19e6e9865e45e0ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/793798
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Marina Ciocea <marinaciocea@google.com>