[starnix][magma] Filter exposed devices.
At the moment only one device can be exposed to the starnix guest, so
filter devices based on an allowlist so only useful devices are exposed,
rather than just the randomly-first device on a system.

Also track magma_device_t and zircon channel lifetimes a bit better.
magma_device_import takes ownership of the zircon channel, so we should
be attempting to store it in an RAII type, rather than storing the
channel.

Bug: 100454
Change-Id: Ie66051f7c0be9cc096e6068e36cc548017368b65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857285
Reviewed-by: Craig Stout <cstout@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: John Bauman <jbauman@google.com>