[assembly] Enable platform AIB for audio
Configure the media subsystem for software assembly. Currently,
we only enable the 'audio_dev_support' AIB (containing the
'virtual_audio' driver and eventually the 'ffx audio' daemon).
We do this on the 'eng' build-type and 'minimal' feature-set.

Thus it is included in (for example):
    core_with_netstack3, terminal, workstation, smart_display_eng
but not in (for example):
    bringup, smart_display_userdebug

Change-Id: I28806334d071b999f055c6b162ad48a541c017c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/841645
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Cam Lloyd <camlloyd@google.com>