[starnix] Refactor file offset handling
File offset is now completely owned and modified by FileOps. File
implementation specify whether the file is seekable or not (and whether
the offset are meaningful or not for file that accepts seek but have not
notion of file position), and FileObject update the offset depending on
result of read/write when needed.
This allows FileOperation to have a single operation that always pass an
offset. This offset will always be 0 for file that do not have
meaningful offsets.

Change-Id: Ic2d1cd23cbf021ac4dad193d216c88704ae148b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857759
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Barth <abarth@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>