[starnix] Remove exception executor
The restricted executor has now reached feature parity with the
exception executor, and is significantly faster.

Maintaining two different executors is a maintenance burden: right now
the exception executor is not even compiled in CQ/CI.

Given that Zircon considers restricted mode part of the kernel's base
functionality, and is committed to supporting it on new architectures
moving forward, the exception executor is unlikely to be useful in the
future.

Change-Id: Ia510a9243060ee160cbdd0d7a8ac23f008d77d28
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/849157
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>