[starnix] Extract __kernel_rt_sigreturn offset in vDSO
Compute the offset of sigreturn in the vDSO and store it alongside the
vmo containing the implementation.
Also store the address at which the VvDSO is mapped on the memory
manager and then use this address together with the offset for signal
trampoline address computation.

Change-Id: I70ce7231b28129fcd1743e6af5471f9b22d43d47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851717
Reviewed-by: Benjamin Lerman <qsr@google.com>
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Maria Glukhova <mariagl@google.com>