[ffx] daemon stop can fail in self-test
Removed the use of "ffx daemon stop" as a way of cleaning up
at the end of tests, since just removing the Isolate is
more reliable.

.. or it should be, anyway.  In the mac it doesn't seem to be.

Even when we _do_ invoke "daemon stop", we never want to use "-w"
in a test since that will make it unclear when the shutdown is
taking unexpectedly long.

Bug: 125459
Fixed: 127648
Change-Id: I089c9ccf7c05394cc035f49cd25d6a1e4ad52e84
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858956
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>