[assembly][abi_revison validation] Remove unused allowlist mechanism for flag to downgrade errors to warnings
Switch from using an allowlist of package names that are downgraded
to a warning from an error if they fail validation, to a single
option that applies to all abi_revision validation errors.  This
flag currently defaults to warning on abi_revisions validation
errors, and will soon be transitioned to defaulting to being an
error when the remaining failures are addressed.

Bug: 91001
Change-Id: I1fc3776794b2508b075f7b6ecbb1ad3564944a51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859887
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>