[cml] Set rules for aggregation
'Aggregation' is the act of routing multiple capabilities to the same
target with the same name, creating (in effect) an aggregate capability.

Example:

```
expose: [
    {
        service: "fuchsia.examples.Echo",
        from: [ "#boot-drivers", "#pkg-drivers" ],
    },
],
collections: [
    {
        name: "boot-drivers",
        durability: "transient",
    },
    {
        name: "pkg-drivers",
        durability: "transient",
    },
],
```

Enforce rules for when aggregation may be used in cml, which is (for
now):

- The capability must be a service
- The routing directive is `expose` or `offer`
- All sources must belong to the same `from` grouping
- All sources are collections

Later, we will relax these rules to allow children to be sources, once
runtime support for that type of aggregation is landed.

Fixed: 127502
Bug: 4776
Change-Id: I533ba09c12cad044e45047b8aad6070522d9e2e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859882
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Gary Bressler <geb@google.com>