[terminal] Add log encoding validator package
This is necessary as part of the effort to move Dart stuff out of
fuchsia.git. We still need conformance testing coverage for the Dart
implementation of logs and inspect. Until RFC-0208
(distributing packages in the sdk) is implemented we'll be placing
the necessary packages in the terminal image to run our tests OOT.

Bug: b/240154194
Bug: 127138
Change-Id: I3af68e6da459b689478d35acc3fd17c600ffc4ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/860317
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>