[remote-control] Support moniker maps on ConnectCapability
This brings the equivalent of RemoteControl/Connect "selector maps" to
ConnectCapability in the form of moniker maps. For now the moniker map
is just a hashamp. Selector maps supported recursive mppaings, for now
moniker maps don't. If we need that in the future we can add it.

The goal of the map is to provide compatibility with older tools when
the moniker in the topology changes. This is aligned with RFC-0169 which
requires newer products to be compatible with older tools for the
duration of the ffx compatibility window.

Just like the current selector map, the moniker map introduced in this
CL is empty.

When we no longer need to support RemoteControl/Connect we can delete
the selector map.

Bug: 126721

Change-Id: Ibcee2a6deadd1b1720aaf3aba948108ab40e2508
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857176
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>