[build][icu] Set the flavored crate names explicitly
The earlier approach didn't set the crate names properly,
causing the compiler to complain about the presence of the `.` in
the crate name.

You may notice that the GN built-in `rust_library` template, and
Fuchsia bespoke `rustc_library` use different parameter names
to specify the crate naming. The former uses `crate_name`, while
the latter uses `name`. This is a bit of an inconsistency, but
probably not worth fixing on a large scale.

Bug: 117974
Change-Id: Ife6f2451c2179f3d2eacb91529d66b9cba954ef7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/860725
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>