[ffx][log] Report counts of dropped and rolled out logs
Looks like this, appended to the end of the log line, in yellow:
[dropped=30202].

This change also ensures that no rolled=N log warnings are generated
when we first read logs from the buffer. It's expected that logs have
been rolled out from the buffer by the time we read and we don't care
about how many. However knowing how many are rolled out as we are
ingesting logs could be useful.

If this information turns out to be too nosiy in the log, we'll add a
flag to optionally add these annotations instead of always adding
them.

Fixed: 127297
Change-Id: I854258a07f590d9a1e7a36fb2953d370659325a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/856595
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Theodore Dubois <tbodt@google.com>