[ffx][isolate] Use normal ffx suite test_data
Changes the e2e_emu library to go through the normal ffx test_suite_data
host test data definition rather than bringing in the SDK.

It also changes the api of `Isolate::new_in_test` a bit to allow for
tests that need the sdk present, and to not require the tests to find
and name the build root directory which is almost always just "." in
test runs. It then derives and configures a reasonable location for the
sdk path in there, using the exported core sdk.

Fixed: 127762

Change-Id: Icfe16d096e123b015d263f7a9b65f12bdd3b2fe8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/860776
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>