[banjo][display] Config stamp values must be strictly increasing.
Previously, the display controller `config_stamp` values in
ApplyConfiguration() are only required to be unique but not
necessarily increasing. Drivers need to do extra mapping from
config stamp values to their chronological order.

This change adds an extra constraints for config_stamp field in
ApplyConfiguration(); config_stamp must be strictly increasing
so that it can stand for the chronological order of a display
configuration.

Note that the display coordinator (the only client of the banjo
display protocol) is already following this requirements, so this
API change is compatible with the existing client implementation.

Test: API-change only. No expected behavior change.
Bug: 126254
Change-Id: I0d28d940efabe9b89ed555d469b571140d334be2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/858890
Reviewed-by: John Bauman <jbauman@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
API-Review: John Bauman <jbauman@google.com>