[radar][proxy] Split radar-proxy into two binaries for injection
Now that there are two AIBs for radar-proxy, the binary can be split so
that the provider proxying and injection code is not included on user
builds. This also requires splitting the test -- shared utilities have
been moved to a header, and test cases to separate source files.

Bug: 100020
Test: fx test radar-integration-test radar-proxy-test
Test: Radar features work on Nelson eng and user builds
Change-Id: Ifee64676f9782fa3df110511ce0e04e1611647e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/859876
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>