[netstack3] Expose testutils with feature
When the "testutils" feature is enabled, expose some core testutil types
so they can be used in other crates. This allows other crates to more
easily test against a fake SyncCtx.

To demonstrate the utility of this approach, make the fuzzing binary an
external consumer of Netstack3 core instead of building it as another
version of the crate. This reduces the space of configurations that
netstack3-core is required to support by removing special handling for
fuzzing.

Bug: 125654
Change-Id: I295d4f92d3c68fd059a5c038dfa8ceb6638ffa2f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/857418
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>