[fxfs] Only examine golden files from the current build.
Previously we would list the directory and run the golden tests against
all files found with a .zstd extension, but since files added in
previous builds do not get removed in future builds this means that we
may examine files that are not meant to be supported by the subsequent
builds.

Now we add a json manifest file to track exactly which golden images are
expected for the current build and only examine those.

Fix: 99892
Change-Id: I21f1da5a44742177780b5b85ef2a6dd7cfa8d0e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862219
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>