[ffx][e2e_emu] Run a package server.
Allows developers to iterate on e2e tests locally when their
test's package deps are in the universe set.

To be meaningful, the test for this functionality needs to
resolve a package which is not in the system image produced
by the main build graph. There is a "unit" test in ffx_e2e_emu
for both the successful case with a package server running and
a failure case without one. This ensures that the package is not
available without a package server and is thus in the "universe"
package set as if a developer had added it to their build with
`fx set ... --with //foo`.

Change-Id: Ib521c93b4ce3dd66ede4a98920265b3e71ba88e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/855101
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Steven Grady <slgrady@google.com>