[display-png] Remove gamma table support.
The display API is not supporting setting gamma table
anymore, and we are going to remove this from carnelian
and all the clients that uses this feature.

This removes all the gamma table related code and tests from
display-png binary tool.

Bug: 127817
Test: display_png --background ff0000ff
(Note that the test is failing on ToT when displaying PNG files,
this is tracked in Bug 127903.)

Change-Id: I029dfdef11738793bc75c6440bb7e142a2840cb5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862476
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Victor Costan <costan@google.com>