[pkg-resolver] Log beginning of ephemeral resolve attempt
Devs use this log to determine if an ongoing resolve is the reason
for the device taking a while to do something.

Resolve state is available in inspect, but checking inspect is not
part of the common dev workflow and it's not common knowledge that
pkg-resolver inspect state should be checked for this information.

Change-Id: Id53747339b85387018d720d20bafcbe033177151
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862152
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Reviewed-by: Sen Jiang <senj@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>