[fake-display] Rename MockDisplayDeviceTree -> FakeDisplayStack.
MockDisplayDeviceTree was created from the original (now deleted)
FakeDisplayDeviceTree in fxrev.dev/756652 due to fake_ddk -> mock_ddk
migration.

However, this naming may cause some confusions: As described in [1],
fakes are lightweight implementations of APIs that behaves similar to
the real implementation, while mocks usually stand for interactive
tests to validate functions are called in the correct way.

mock-ddk framework [2] is actually a framework which supports both
faking (it provides a fake device root and the whole device tree
implementation) and mocking (it provides interfaces for tests to
verify interactions with the driver host).

For the MockDisplayDeviceTree class, only the zx_device instance
(mock_root, mock_display) service provision is mocked so that they
can provide services from certain FIDL / banjo servers;
the display device on the device tree is a fake but not a mock.
Thus, this change replaces the "Mock" in the class name with "Fake"
instead.

Besides, the "FakeDisplayDeviceTree" would cause the confusion that
the class is a fake implementation of a DisplayDeviceTree / DeviceTree
interface; and the device tree itself is an implementation detail which
is not exposed to the users. In fact, the topology among display
driver, display coordinator and sysmem doesn't matter and may subject
to change with the driver framework. Thus, this change also replaces
the "DeviceTree" in the class name with "Stack" instead; clients may
see a bag of devices but they don't need to make any assumptions about
the device topology.

[1] https://abseil.io/resources/swe-book/html/ch13.html#techniques_for_using_test_doubles
[2] //docs/development/drivers/testing/mock_ddk.md

Test: no-op.
Change-Id: Idb6de7fc26e33ae51cbba6c713705913ca3de272
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862142
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Victor Costan <costan@google.com>